@@ -6,16 +6,16 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.0+2
-  url_launcher: ^5.4.0
-  clipboard_manager: ^0.0.1
+  http: ^0.13.4
+  url_launcher: ^6.0.20
 
-  cupertino_icons: ^0.1.3
+  # cupertino_icons: ^1.0.4
+  google_fonts: ^2.3.1
 
 dev_dependencies:
   flutter_test:
@@ -31,4 +31,4 @@ flutter:
   fonts:
     - family: GoogleSansRegular
       fonts:
-      - asset: assets/GoogleSansRegular.ttf
+      - asset: assets/fonts/GoogleSansRegular.ttf
