@@ -4,6 +4,6 @@ class Project{
   String image;
   String name;
   String description;
-  String link;
-  Project({@required this.image,@required this.name,@required this.description, this.link});
+  String? link;
+  Project({required this.image,required this.name,required this.description, this.link});
 }
